Move the Tx protocol into the database package
serviceshared/tx.py only existed to break the import cycle between the
database package init and the matching models. Deferring the tracker
import to TxCursor construction (a sys.modules hit after the first call)
breaks the cycle just as well, so the protocol can live where it belongs,
as serviceshared/database/tx.py.
